Add few fixes after review
