+++ frontend/src/features/map/layers/styles/utils.test.ts
@@ -3,11 +3,43 @@
 import { describe, expect, it } from '@jest/globals'
 import dayjs from 'dayjs'
 
-import { getColorWithAlpha, getStartAndEndDatesSetWithCurrentYear } from './utils'
+import { getColorWithAlpha, getHashDigitsFromRegulation, getStartAndEndDatesSetWithCurrentYear } from './utils'
 import { theme } from '../../../../ui/theme'
 import { getUtcDayjs } from '../../../../utils/getUtcDayjs'
 
 describe('utils', () => {
+  it('getHashDigitsFromRegulation Should return undefined When the regulation is null', async () => {
+    // Given
+    const regulation = null
+
+    // When
+    expect(getHashDigitsFromRegulation(regulation)).toBeUndefined()
+  })
+
+  it('getHashDigitsFromRegulation Should return a constant number for a given regulatory zone', async () => {
+    // Given
+    const regulation = {
+      topic: 'Ouest Cotentin Bivalves',
+      zone: 'Zone ouest'
+    }
+
+    // Then
+    expect(getHashDigitsFromRegulation(regulation)).toEqual(11)
+    // Retry to ensure the returned digit is constant
+    expect(getHashDigitsFromRegulation(regulation)).toEqual(11)
+  })
+
+  it('getHashDigitsFromRegulation Should return a constant number for another regulatory zone', async () => {
+    // Given
+    const regulation = {
+      topic: 'Ouest Cotentin Bivalves',
+      zone: 'Zone nord'
+    }
+
+    // When
+    expect(getHashDigitsFromRegulation(regulation)).toEqual(9)
+  })
+
   it('getColorWithAlpha Should add the alpha field on a color', async () => {
     // Given
     const color = theme.color.lightBlue
